requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2

# Screenshot and image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# OCR and Computer Vision
pytesseract==0.3.10
ultralytics==8.0.196

# Voice and Audio - FREE ALTERNATIVES
SpeechRecognition==3.10.0
pyaudio==0.2.11
openai-whisper==20231117  # Free local Whisper
pyttsx3==2.90  # Free local TTS
sounddevice==0.4.6  # Better audio handling

# Screen automation
pyautogui==0.9.54
keyboard==0.13.5
mouse==0.7.1

# AI and LLM - FREE ALTERNATIVES
ollama==0.1.7  # Free local LLM client
transformers==4.35.0  # Hugging Face transformers
torch==2.1.0  # PyTorch for local models

# Local storage - FREE ALTERNATIVES
sqlite3  # Built into Python
aiosqlite==0.19.0  # Async SQLite

# GUI Framework - MODERN UI (Choose one based on your system)
# Option 1: PyQt6 (recommended for newer systems)
# PyQt6>=6.4.0
# Option 2: PyQt5 (better compatibility)
# PyQt5>=5.15.0
# Option 3: Tkinter (built into Python, most compatible)
# tkinter  # Built into Python

# Web framework for API (optional)
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6 